refactor: split ItemList.js into multiple components
