+++ scripts/apps/search/components/ErrorBox.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export function ErrorBox(props) {
+    const {gettextCatalog} = props.svc;
+
+    return React.createElement('div', {className: 'error-box'},
+        React.createElement('p', {className: 'message'},
+        gettextCatalog.getString('There was an error archiving this item')),
+        React.createElement('div', {className: 'buttons'})
+    );
+}
+
+ErrorBox.propTypes = {
+    svc: React.PropTypes.object.isRequired
+};
