+++ scripts/apps/search/components/ProgressBar.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export function ProgressBar(props) {
+    return React.createElement('div', {
+        className: 'archiving-progress',
+        style: {width: props.completed + '%'}
+    });
+}
+
+ProgressBar.propTypes = {
+    completed: React.PropTypes.any
+};
