+++ scripts/apps/search/components/actions-menu/Divider.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Divider() {
+    return React.createElement('li', {className: 'divider'});
+}
