+++ scripts/apps/search/components/actions-menu/index.js
@@ -1,0 +1,1 @@
+export {ActionsMenu} from './ActionsMenu';
