+++ scripts/apps/search/components/fields/index.js
@@ -1,0 +1,23 @@
+export {headline} from './headline';
+export {slugline} from './slugline';
+export {wordcount} from './wordcount';
+export {highlights} from './highlights';
+export {versioncreated} from './versioncreated';
+export {versioncreator} from './versioncreator';
+export {profile} from './profile';
+export {state} from './state';
+export {embargo} from './embargo';
+export {update} from './update';
+export {updated} from './updated';
+export {takekey} from './takekey';
+export {signal} from './signal';
+export {broadcast} from './broadcast';
+export {flags} from './flags';
+export {takepackage} from './takepackage';
+export {category} from './category';
+export {provider} from './provider';
+export {expiry} from './expiry';
+export {desk} from './desk';
+export {fetchedDesk} from './fetchedDesk';
+export {priority} from './priority';
+export {urgency} from './urgency';
