+++ scripts/apps/search/components/fields/provider.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export function provider(props) {
+    var provider = props.ingestProvider ? props.ingestProvider.source : '';
+
+    if (props.item.source) {
+        provider = props.item.source;
+    }
+    if (provider) {
+        return React.createElement('span', {className: 'provider', key: 'provider'}, provider);
+    }
+}
+
+provider.propTypes = {
+    item: React.PropTypes.any,
+    ingestProvider: React.PropTypes.any,
+};
