+++ scripts/apps/search/components/fields/signal.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export function signal(props) {
+    if (props.item.signal) {
+        return React.createElement('span', {className: 'signal', key: 'signal'}, props.item.signal);
+    }
+}
+
+signal.propTypes = {
+    item: React.PropTypes.any,
+};
