+++ scripts/apps/search/components/fields/versioncreated.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import {TimeElem} from 'apps/search/components';
+
+export function versioncreated(props) {
+    return React.createElement(
+        TimeElem, {
+            date: props.item.versioncreated,
+            key: 'versioncreated',
+            svc: props.svc
+        }
+    );
+}
+
+versioncreated.propTypes = {
+    svc: React.PropTypes.object.isRequired,
+    item: React.PropTypes.any,
+};
