+++ scripts/apps/search/components/fields/wordcount.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export function wordcount(props) {
+    return React.createElement(
+        'span',
+        {className: 'word-count', key: 'wordcount'},
+        props.item.word_count
+    );
+}
+
+wordcount.propTypes = {
+    item: React.PropTypes.any,
+};
