+++ scripts/apps/search/directives/ItemList.js
@@ -1,14 +1,10 @@
-/* eslint-disable react/prop-types */
 /* eslint-disable react/no-render-return-value */
-/* eslint-disable react/prefer-es6-class */
-/* eslint-disable react/no-multi-comp */
-
-// TODO(*): refactor this to use JSX and multiple files
+// TODO(*): Fix above?
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import classNames from 'classnames';
-import {DEFAULT_LIST_CONFIG} from 'apps/search/constants';
+
+import {ItemList as ItemListComponent} from 'apps/search/components';
 
 ItemList.$inject = [
     '$location',
@@ -67,90 +63,40 @@ export function ItemList(
     metadata,
     storage
 ) {
-    var listConfig = config.list || DEFAULT_LIST_CONFIG;
-
-    function getSpecStyle(spec) {
-        var style = {};
-
-        if (spec.color) {
-            style.backgroundColor = spec.color;
-        }
-
-        return style;
-    }
-
-    function getSpecTitle(spec, title) {
-        return spec.name || title;
-    }
-
-    function getSpecValue(spec, value) {
-        return spec.short || value;
-    }
-
-    var ItemPriority = function(props) {
-        var priority = props.priority || 3;
-        var spec = metadata.priorityByValue(priority);
-
-        if (spec) {
-            return React.createElement(
-                'span',
-                {
-                    className: 'priority-label priority-label--' + priority,
-                    style: getSpecStyle(spec),
-                    title: getSpecTitle(spec, gettext('Priority')),
-                    key: 'priority'
-                },
-                getSpecValue(spec, priority)
-            );
-        }
-
-        return React.createElement(
-            'span',
-            {className: 'priority-label priority-label--' + priority, title: gettext('Priority'), key: 'priority'},
-            priority
-        );
-    };
-
-    var ItemUrgency = function(props) {
-        var urgency = props.urgency || 3;
-        var spec = metadata.urgencyByValue(urgency);
-
-        if (spec) {
-            return React.createElement(
-                'span',
-                {
-                    className: 'urgency-label urgency-label--' + urgency,
-                    title: getSpecTitle(spec, gettextCatalog.getString('Urgency')),
-                    style: getSpecStyle(spec),
-                    key: 'urgency'
-                },
-                getSpecValue(spec, urgency)
-            );
-        }
-
-        return React.createElement(
-            'span',
-            {
-                className: 'urgency-label urgency-label--' + urgency,
-                title: gettextCatalog.getString('Urgency'),
-                key: 'urgency'
-            },
-            urgency
-        );
+    // contains all the injected services to be passed down to child
+    // components via props
+    const services = {
+        $location: $location,
+        $timeout: $timeout,
+        $injector: $injector,
+        $filter: $filter,
+        search: search,
+        datetime: datetime,
+        gettext: gettext,
+        superdesk: superdesk,
+        workflowService: workflowService,
+        archiveService: archiveService,
+        activityService: activityService,
+        multi: multi,
+        desks: desks,
+        familyService: familyService,
+        Keys: Keys,
+        dragitem: dragitem,
+        highlightsService: highlightsService,
+        TranslationService: TranslationService,
+        monitoringState: monitoringState,
+        authoringWorkspace: authoringWorkspace,
+        gettextCatalog: gettextCatalog,
+        $rootScope: $rootScope,
+        config: config,
+        $interpolate: $interpolate,
+        metadata: metadata,
+        storage: storage
     };
 
     return {
         link: function(scope, elem) {
             elem.attr('tabindex', 0);
-
-            var menuHolderElem = document.getElementById('react-placeholder');
-            var closeActionsMenu = function() {
-                ReactDOM.unmountComponentAtNode(menuHolderElem);
-            };
-
-            var getViewType = function() {
-                return scope.viewType;
-            };
 
             var groupId = scope.$id;
             var groups = monitoringState.state.groups || [];
@@ -166,1621 +112,13 @@ export function ItemList(
                 }
             });
 
-            /**
-             * Render element close to target, but append to body
-             *
-             * Used for dropdown menus that would be only partialy visible if rendered
-             * within parent which has overflow: hidden that is required for scrolling
-             *
-             * @param {Object} elem React element
-             * @param {Node} target DOM node
-             */
-            function renderToBody(elem, target) {
-                // first render it somewhere not visible
-                menuHolderElem.style.zIndex = -1;
-                var node = ReactDOM.findDOMNode(ReactDOM.render(elem, menuHolderElem));
-                // make sure it's rendered
-
-                node.style.display = 'block';
-                var rect = node.getBoundingClientRect();
-                var width = rect.width;
-                var height = rect.height;
-
-                var ACTION_MENU_FROM_TOP = 150; // 150 = top-menu + search bar
-                var MENU_MARGIN_HEIGHT = 16;
-                var LEFT_BAR_WIDTH = 48;
-                var BOTTOM_BAR_HEIGHT = 30;
-                var BUFFER = 5;
-
-                // get target position
-                var targetRect = target.getBoundingClientRect();
-
-                // get workspace
-                var workspace = document.getElementById('main-container');
-
-                // compute menu position
-                var top = targetRect.top + targetRect.height;
-                var left = targetRect.left + targetRect.width - width;
-
-                // menu goes off on the right side
-                if (left + width + BUFFER > workspace.clientWidth) {
-                    left -= width;
-                    left += targetRect.width;
-                }
-
-                // menu goes off on the left side
-                if (left - LEFT_BAR_WIDTH < 0) {
-                    left = targetRect.left;
-                }
-
-                // menu goes out on the bottom side
-                if (top + height + BOTTOM_BAR_HEIGHT + BUFFER > workspace.clientHeight) {
-                    top -= height;
-                    top -= targetRect.height;
-                    top -= MENU_MARGIN_HEIGHT;
-                    top = top < ACTION_MENU_FROM_TOP ? ACTION_MENU_FROM_TOP : top;
-                }
-
-                node.style.top = top.toFixed() + 'px';
-                node.style.left = left.toFixed() + 'px';
-                node.style.position = 'absolute';
-                menuHolderElem.style.zIndex = 1000;
-            }
-
-            /**
-             * Test if an item has thumbnail
-             *
-             * @return {Boolean}
-             */
-            function hasThumbnail(item) {
-                return item.renditions && item.renditions.thumbnail;
-            }
-
-            /**
-             * Test if an multi-selection allowed
-             *
-             * @return {Boolean}
-             */
-            function isCheckAllowed(item) {
-                return !(item.state === 'killed' || item._type === 'published' && !item.last_published_version);
-            }
-
-            /**
-             * Time element
-             */
-            var TimeElem = function(props) {
-                return React.createElement(
-                    'time',
-                    {title: datetime.longFormat(props.date)},
-                    datetime.shortFormat(props.date)
-                );
-            };
-
-            function createMarkUp(html) {
-                return {__html: html};
-            }
-
-            /**
-             * Media Preview - renders item thumbnail
-             */
-            var MediaPreview = function(props) {
-                var item = props.item;
-                var headline = item.headline || item.slugline || item.type;
-                // headline could contains html tags hence stripping for tooltips
-                var headlineText = headline.replace(/(<([^>]+)>)/ig, '');
-                var preview;
-
-                if (hasThumbnail(props.item)) {
-                    preview = React.createElement(
-                        'img',
-                        {src: props.item.renditions.thumbnail.href}
-                    );
-                }
-
-                return React.createElement(
-                    'div',
-                    {className: 'media multi'},
-                    preview ? React.createElement(
-                        'figure',
-                        null,
-                        preview
-                    ) : null,
-                    React.createElement(
-                        'span',
-                        {className: 'text'},
-                        React.createElement(
-                            'small',
-                            {title: headlineText,
-                                dangerouslySetInnerHTML: createMarkUp(headline)}
-                        ),
-                        React.createElement(ItemContainer, {item: item, desk: props.desk})
-                    ),
-                    React.createElement(SelectBox, {item: item, onMultiSelect: props.onMultiSelect})
-                );
-            };
-
-            var SelectBox = React.createClass({
-                toggle: function(event) {
-                    event.stopPropagation();
-                    if (isCheckAllowed(this.props.item)) {
-                        var selected = !this.props.item.selected;
-
-                        this.props.onMultiSelect([this.props.item], selected);
-                    }
-                },
-
-                render: function() {
-                    if (this.props.item.selected) {
-                        this.props.item.selected = isCheckAllowed(this.props.item);
-                    }
-                    return React.createElement(
-                        'div',
-                        {
-                            className: 'selectbox',
-                            title: isCheckAllowed(this.props.item) ? null : 'selection not allowed',
-                            onClick: this.toggle
-                        },
-                        React.createElement(
-                            'span', {
-                                className: 'sd-checkbox' + (this.props.item.selected ? ' checked' : '')
-                            }
-                        )
-                    );
-                }
-            });
-
-            var ItemContainer = function(props) {
-                var item = props.item;
-                var desk = props.desk || null;
-                var label, value;
-
-                if (item._type !== 'ingest') {
-                    if (desk) {
-                        label = gettext('desk:');
-                        value = desk.name;
-                    } else if (item._type === 'archive') {
-                        label = gettext('location:');
-                        value = gettext('workspace');
-                    } else if (item._type === 'archived') {
-                        label = '';
-                        value = gettext('archived');
-                    }
-                }
-
-                return React.createElement(
-                    'span',
-                    {className: 'container', title: value ? label + ' ' + value : null},
-                    React.createElement(
-                        'span',
-                        {className: 'location-desk-label'},
-                        label
-                    ),
-                    value
-                );
-            };
-
-            /**
-             * Media Info - renders item metadata
-             */
-            var MediaInfo = function(props) {
-                var item = props.item;
-                var meta = [];
-                var source = props.ingestProvider ? props.ingestProvider.source : '';
-
-                if (item.source) {
-                    source = item.source;
-                }
-                if (source) {
-                    meta.push(
-                        React.createElement('dt', {key: 1}, gettextCatalog.getString('source')),
-                        React.createElement('dd', {key: 2, className: 'provider'}, source)
-                    );
-                }
-
-                meta.push(
-                    React.createElement('dt', {key: 3}, gettextCatalog.getString('updated')),
-                    React.createElement('dd', {key: 4}, datetime.shortFormat(item.versioncreated))
-                );
-
-                if (item.is_spiked) {
-                    meta.push(React.createElement('dt', {key: 5}, gettextCatalog.getString('expires')));
-                    meta.push(React.createElement('dd', {key: 6}, datetime.shortFormat(item.expiry)));
-                }
-
-                var info = [];
-                var flags = item.flags || {};
-
-                info.push(React.createElement(
-                    'h5',
-                    {key: 1,
-                        dangerouslySetInnerHTML: createMarkUp(item.headline || item.slugline || item.type)}
-
-                ));
-
-                info.push(React.createElement(
-                    'dl',
-                    {key: 2},
-                    meta
-                ));
-
-                if (flags.marked_for_legal) {
-                    info.push(React.createElement(
-                        'div',
-                        {key: 3, className: 'state-label legal'},
-                        gettext('Legal')
-                    ));
-                }
-
-                if (item.archived) {
-                    info.push(React.createElement(
-                        'div',
-                        {className: 'fetched-desk', key: 4},
-                        React.createElement(FetchedDesksInfo, {item: item})
-                    ));
-                }
-
-                return React.createElement('div', {className: 'media-info'}, info);
-            };
-
-            /**
-             * Type icon component
-             */
-            var TypeIcon = function(props) {
-                if (props.package_type) {
-                    return React.createElement('i', {className: 'filetype-icon-takes-pack'});
-                } else if (props.type === 'composite' && props.highlight) {
-                    return React.createElement('i', {className: 'filetype-icon-highlight-pack'});
-                }
-
-                return React.createElement('i', {className: 'filetype-icon-' + props.type});
-            };
-
-            var GridTypeIcon = function(props) {
-                return React.createElement(
-                    'span',
-                    {className: classNames('type-icon', {swimlane: props.swimlane})},
-                    React.createElement(TypeIcon, {type: props.item.type})
-                );
-            };
-
-            var ListTypeIcon = React.createClass({
-                getInitialState: function() {
-                    return {hover: false};
-                },
-
-                setHover: function() {
-                    this.setState({hover: true});
-                },
-
-                unsetHover: function() {
-                    if (this.state.hover) {
-                        this.setState({hover: false});
-                    }
-                },
-
-                render: function() {
-                    var showSelect = this.state.hover || this.props.item.selected;
-
-                    return React.createElement(
-                        'div',
-                        {
-                            className: classNames('list-field type-icon'),
-                            onMouseEnter: this.setHover,
-                            onMouseLeave: this.unsetHover
-                        },
-                        showSelect ?
-                            React.createElement(SelectBox, {
-                                item: this.props.item,
-                                onMultiSelect: this.props.onMultiSelect
-                            }) :
-                            React.createElement(
-                                TypeIcon,
-                                {
-                                    type: this.props.item.type,
-                                    package_type: this.props.item.package_type,
-                                    highlight: this.props.item.highlight
-                                }
-                            )
-                    );
-                }
-            });
-
-            var ListPriority = function(props) {
-                var css = {
-                    className: classNames('list-field urgency', {
-                        'urgency-reduced-rowheight': listConfig.thinRows
-                    })
-                };
-
-                return renderArea('priority', props, css) || React.createElement('div', css);
-            };
-
-            var HighlightsList = React.createClass({
-                removeHighlight: function(highlight) {
-                    return function(event) {
-                        event.stopPropagation();
-                        highlightsService.markItem(highlight._id, this.props.item);
-
-                        if (getViewType() === 'highlights' && this.props.item.highlights.length === 1) {
-                            $rootScope.$broadcast('multi:remove', this.props.item._id);
-                        }
-                    }.bind(this);
-                },
-
-                componentDidMount: function() {
-                    this.timeout = null;
-                },
-
-                componentWillUnmount: function() {
-                    this.stopTimeout();
-                },
-
-                stopTimeout: function() {
-                    this.timeout = $timeout.cancel(this.timeout);
-                },
-
-                close: function() {
-                    this.timeout = $timeout(closeActionsMenu, 2000, false);
-                },
-
-                render: function() {
-                    var highlights = this.props.highlights;
-                    var highlightsById = this.props.highlightsById || {};
-
-                    var createHighlight = function(id) {
-                        var highlight = highlightsById[id];
-
-                        if (highlight) {
-                            return React.createElement(
-                                'li',
-                                {key: 'item-highlight-' + highlight._id},
-                                highlight.name,
-                                highlightsService.hasMarkItemPrivilege() ? React.createElement(
-                                    'button',
-                                    {className: 'btn btn--mini', onClick: this.removeHighlight(highlight)},
-                                    gettext('REMOVE')
-                                ) : null
-                            );
-                        }
-                    }.bind(this);
-
-                    var items = [
-                        React.createElement(
-                            'li',
-                            {key: 'item-highlight-label'},
-                            React.createElement(
-                                'div',
-                                {className: 'dropdown__menu-label'},
-                                gettext('Marked For')
-                            ),
-                            React.createElement(
-                                'button',
-                                {className: 'dropdown__menu-close', onClick: closeActionsMenu},
-                                React.createElement(
-                                    'i',
-                                    {className: 'icon-close-small icon-white'}
-                                )
-                            )
-                        )
-                    ];
-
-                    return React.createElement(
-                        'ul',
-                        {
-                            className: 'dropdown dropdown__menu highlights-list-menu open',
-                            onMouseEnter: this.stopTimeout,
-                            onMouseLeave: this.close
-                        },
-                        items.concat(highlights.map(createHighlight))
-                    );
-                }
-            });
-
-            var HighlightsInfo = React.createClass({
-                toggle: function(event) {
-                    if (event) {
-                        event.stopPropagation();
-                    }
-
-                    closeActionsMenu();
-                    this.renderDropdown();
-                },
-
-                /**
-                 * Checks if the given item is in the daterange of the highlights
-                 * for every highlight given
-                 *
-                 * @param {array} highlights
-                 * @param {Object} item
-                 * @return {Object}
-                 */
-                getHighlightStatuses: function(highlights, item) {
-                    var highlightStatuses = {};
-                    var highlightsById = this.props.highlightsById;
-
-                    _.forEach(highlights, (highlight) => {
-                        var hours = $filter('hoursFromNow')(item.versioncreated);
-
-                        highlightStatuses[highlight] = highlightsService.isInDateRange(
-                            highlightsById[highlight], hours
-                        );
-                    });
-
-                    return highlightStatuses;
-                },
-
-                getHighlights: function() {
-                    var itemHighlights = [];
-
-                    if (isCheckAllowed(this.props.item)) {
-                        if (this.props.item.archive_item && this.props.item.archive_item.highlights &&
-                            this.props.item.archive_item.highlights.length) {
-                            itemHighlights = this.props.item.archive_item.highlights;
-                        } else {
-                            itemHighlights = this.props.item.highlights || [];
-                        }
-                    }
-
-                    return itemHighlights;
-                },
-
-                render: function() {
-                    var highlights = this.getHighlights();
-
-                    var hasActiveHighlight = function() {
-                        var statuses = this.getHighlightStatuses(highlights, this.props.item);
-
-                        if ($location.path() === '/workspace/highlights') {
-                            return statuses[$location.search().highlight];
-                        }
-
-                        return highlights.some((h) => statuses[h]);
-                    }.call(this);
-
-                    return React.createElement(
-                        'div',
-                        {
-                            className: 'highlights-box',
-                            onClick: this.toggle
-                        },
-                        highlights.length ? React.createElement(
-                            'div',
-                            {className: 'highlights-list dropdown'},
-                            React.createElement(
-                                'button',
-                                {className: 'dropdown__toggle'},
-                                React.createElement('i', {
-                                    className: classNames({
-                                        'icon-star': highlights.length === 1,
-                                        'icon-multi-star': highlights.length > 1,
-                                        red: hasActiveHighlight
-                                    })
-                                })
-                            )
-                        ) : null
-                    );
-                },
-
-                renderDropdown: function() {
-                    var elem = React.createElement(HighlightsList, {
-                        item: this.props.item,
-                        highlights: this.getHighlights(),
-                        highlightsById: this.props.highlightsById
-                    });
-
-                    var icon = ReactDOM.findDOMNode(this).getElementsByClassName('icon-star')[0] ||
-                    ReactDOM.findDOMNode(this).getElementsByClassName('icon-multi-star')[0];
-
-                    renderToBody(elem, icon);
-                }
-            });
-
-            var closeTimeout;
-            var DesksDropdown = React.createClass({
-                getInitialState: function() {
-                    return {open: false};
-                },
-
-                close: function(cancel) {
-                    if (cancel === true && closeTimeout) {
-                        $timeout.cancel(closeTimeout);
-                    } else {
-                        closeTimeout = $timeout(() => {
-                            closeActionsMenu();
-                        }, 200, false);
-                    }
-                },
-
-                toggle: function(event) {
-                    if (event) {
-                        event.stopPropagation();
-                    }
-
-                    this.close(true);
-                    this.renderDropdown();
-                },
-
-                render: function() {
-                    return React.createElement('dd',
-                            {className: 'dropdown dropdown--dropup more-actions'},
-                            React.createElement('button', {
-                                className: 'dropdown__toggle',
-                                onMouseOver: this.toggle,
-                                onMouseLeave: this.close
-                            },
-                            React.createElement('i',
-                                    {className: 'icon-dots'})
-                            ));
-                },
-
-                renderDropdown: function() {
-                    var desks = this.props.desks.map((desk, index) => React.createElement(
-                                'li',
-                                {key: 'desk' + index},
-                                React.createElement(
-                                        'a',
-                                        {disabled: !desk.isUserDeskMember, onClick: this.props.openDesk(desk)},
-                                        desk.desk.name + ' (' + desk.count + ')'
-                                        )
-                                ));
-
-                    var elem = React.createElement('div',
-                        {
-                            className: 'dropdown__menu more-activity-menu',
-                            onMouseOver: this.toggle,
-                            onMouseLeave: this.close
-                        },
-                            React.createElement('ul', {}, desks)
-                            );
-
-                    var icon = ReactDOM.findDOMNode(this).getElementsByClassName('dropdown__toggle')[0];
-
-                    renderToBody(elem, icon);
-                }
-            });
-
-            var FetchedDesksInfo = React.createClass({
-                getInitialState: function() {
-                    return {desks: []};
-                },
-
-                componentDidMount: function() {
-                    familyService.fetchDesks(this.props.item, false)
-                        .then((fetchedDesks) => {
-                            this.setState({desks: fetchedDesks});
-                        });
-                },
-
-                componentDidUpdate: function(prevProps, prevState) {
-                    if (prevProps.item !== this.props.item) {
-                        familyService.fetchDesks(this.props.item, false)
-                        .then((fetchedDesks) => {
-                            this.setState({desks: fetchedDesks});
-                        });
-                    }
-                },
-
-                formatDeskName: function(name) {
-                    return name.substr(0, 10) + (name.length > 10 ? '...' : '');
-                },
-
-                openDesk: function(desk) {
-                    return function(event) {
-                        event.stopPropagation();
-                        if (desk.isUserDeskMember) {
-                            desks.setCurrentDeskId(desk.desk._id);
-                            $location.url('/workspace/monitoring');
-                            if (desk.count === 1) {
-                                superdesk.intent('edit', 'item', desk.item);
-                            }
-                        }
-                    };
-                },
-
-                render: function() {
-                    var items = [];
-
-                    items.push(React.createElement('dt', {
-                        key: 'dt',
-                        style: {paddingRight: '5px'}
-                    }, gettext('fetched in')));
-
-                    if (this.state.desks.length) {
-                        var desk = this.state.desks[0];
-                        var name = this.formatDeskName(desk.desk.name);
-
-                        items.push(React.createElement('dd', {key: 'dd1'}, desk.isUserDeskMember ?
-                            React.createElement('a', {onClick: this.openDesk(desk)}, name) :
-                            React.createElement('span', {className: 'container'}, name)
-                        ));
-
-                        if (this.state.desks.length > 1) {
-                            items.push(React.createElement(DesksDropdown, {
-                                key: 'dd2',
-                                desks: this.state.desks,
-                                openDesk: this.openDesk
-                            }));
-                        }
-                    }
-
-                    return React.createElement('div', {},
-                        React.createElement(
-                            'dl',
-                            {},
-                            items
-                        )
-                    );
-                }
-            });
-
-            var ListItemInfo = function(props) {
-                return React.createElement(
-                    'div',
-                    {className: classNames('item-info', {'item-info-reduced-rowheight': listConfig.thinRows})},
-                    renderArea('firstLine', props, {className: 'line'}),
-                    renderArea('secondLine', props, {className: 'line'})
-                );
-            };
-
-            function renderArea(area, itemProps, props) {
-                var specs = listConfig[area] || [];
-                var contents = specs.map((field) => {
-                    if (fields[field]) {
-                        return fields[field](itemProps);
-                    }
-
-                    console.warn('missing field in list: ' + field);
-                    return null;
-                }).filter(angular.identity);
-                var elemProps = angular.extend({key: area}, props);
-
-                return contents.length ? React.createElement('div', elemProps, contents) : null;
-            }
-
-            /**
-             * Fields specified via list config
-             */
-            var fields = {
-                headline: function(props) {
-                    var headline = props.item.headline ? props.item.headline : props.item.type;
-
-                    return React.createElement(
-                        'span',
-                        {className: 'item-heading', key: 'headline',
-                            dangerouslySetInnerHTML: createMarkUp(headline)}
-                    );
-                },
-
-                slugline: function(props) {
-                    if (props.item.slugline) {
-                        return React.createElement(
-                            'span',
-                            {className: 'keyword', key: 'slugline',
-                                dangerouslySetInnerHTML: createMarkUp(props.item.slugline)}
-                        );
-                    }
-                },
-
-                wordcount: function(props) {
-                    return React.createElement(
-                        'span',
-                        {className: 'word-count', key: 'wordcount'},
-                        props.item.word_count
-                    );
-                },
-
-                highlights: function(props) {
-                    return React.createElement(HighlightsInfo, angular.extend({key: 'highlights'}, props));
-                },
-
-                versioncreated: function(props) {
-                    return React.createElement(
-                        TimeElem,
-                        {date: props.item.versioncreated, key: 'versioncreated'}
-                    );
-                },
-
-                versioncreator: function(props) {
-                    return React.createElement(
-                        'span',
-                        {className: 'version-creator', key: 'versioncreator'},
-                        props.versioncreator
-                    );
-                },
-
-                profile: function(props) {
-                    if (props.item.profile) {
-                        return React.createElement(
-                            'div',
-                            {className: 'profile-label profile-label--' + props.item.profile, key: 'profile'},
-                            props.profilesById[props.item.profile] ?
-                                props.profilesById[props.item.profile].label :
-                                props.item.profile
-                        );
-                    }
-                },
-
-                state: function(props) {
-                    return React.createElement(
-                        'span',
-                        {
-                            title: $filter('removeLodash')(props.item.state),
-                            className: 'state-label state-' + props.item.state,
-                            key: 'state'
-                        },
-                        $filter('removeLodash')(gettextCatalog.getString(props.item.state))
-                    );
-                },
-
-                embargo: function(props) {
-                    if (props.item.embargo) {
-                        return React.createElement(
-                            'span',
-                            {className: 'state-label state_embargo', title: gettext('embargo'), key: 'embargo'},
-                            gettext('embargo')
-                        );
-                    }
-                },
-
-                update: function(props) {
-                    if (props.item.correction_sequence) {
-                        return React.createElement(
-                            'div',
-                            {
-                                className: 'provider',
-                                key: 'update',
-                                title: gettextCatalog.getString('correction sequence')
-                            },
-                            $interpolate(
-                                gettextCatalog.getString('Update {{ seq }}'))({
-                                    seq: props.item.correction_sequence
-                                }
-                            )
-                        );
-                    }
-                },
-
-                updated: function(props) {
-                    var selectUpdate = function(event) {
-                        event.stopPropagation();
-                        props.selectUpdate();
-                    };
-
-                    if (props.item.rewritten_by) {
-                        return React.createElement(
-                            'div',
-                            {className: 'state-label updated', key: 'updated', onClick: selectUpdate},
-                            gettextCatalog.getString('Updated')
-                        );
-                    }
-                },
-
-                takekey: function(props) {
-                    if (props.item.anpa_take_key) {
-                        return React.createElement('span', {className: 'takekey', key: 'takekey'},
-                            gettextCatalog.getString(props.item.anpa_take_key));
-                    }
-                },
-
-                signal: function(props) {
-                    if (props.item.signal) {
-                        return React.createElement('span', {className: 'signal', key: 'signal'}, props.item.signal);
-                    }
-                },
-
-                broadcast: function(props) {
-                    var broadcast = props.item.broadcast || {};
-
-                    if (broadcast.status) {
-                        return React.createElement(
-                            'span',
-                            {className: 'broadcast-status', title: broadcast.status, key: 'broadcast'},
-                            '!'
-                        );
-                    }
-                },
-
-                flags: function(props) {
-                    var flags = props.item.flags || {};
-                    var elems = [
-                        flags.marked_for_not_publication ?
-                            React.createElement(
-                                'div', {
-                                    className: 'state-label not-for-publication',
-                                    title: gettext('Not For Publication'),
-                                    key: 'not-for-publication'
-                                }, gettext('Not For Publication'))
-                            : null,
-                        flags.marked_for_legal ?
-                            React.createElement(
-                                'div', {
-                                    className: 'state-label legal',
-                                    title: gettext('Legal'),
-                                    key: 'legal'
-                                }, gettext('Legal'))
-                            : null,
-                        flags.marked_for_sms ?
-                            React.createElement(
-                                'div', {
-                                    className: 'state-label sms',
-                                    key: 'sms'},
-                                gettext('Sms'))
-                            : null
-                    ].filter(angular.identity);
-
-                    return elems.length ? React.createElement('div', {key: 'flags'}, elems) : null;
-                },
-
-                takepackage: function(props) {
-                    var item = props.item;
-                    var archiveItem = props.item.archive_item;
-
-                    var isTake = item.type === 'text' &&
-                        (item.takes && item.takes.sequence > 1 ||
-                        item._type === 'published' && archiveItem &&
-                        archiveItem.takes && archiveItem.takes.sequence > 1);
-
-                    var selectTakesPackage = function(event) {
-                        event.stopPropagation();
-                        props.selectTakesPackage();
-                    };
-
-                    if (isTake) {
-                        return React.createElement(
-                            'div',
-                            {className: 'state-label takes', onClick: selectTakesPackage, key: 'takepackage'},
-                            gettext('Takes')
-                        );
-                    }
-                },
-
-                category: function(props) {
-                    var anpa = props.item.anpa_category || {};
-
-                    if (anpa.name) {
-                        return React.createElement('div', {className: 'category', key: 'category'}, anpa.name);
-                    }
-                },
-
-                provider: function(props) {
-                    var provider = props.ingestProvider ? props.ingestProvider.source : '';
-
-                    if (props.item.source) {
-                        provider = props.item.source;
-                    }
-                    if (provider) {
-                        return React.createElement('span', {className: 'provider', key: 'provider'}, provider);
-                    }
-                },
-
-                expiry: function(props) {
-                    if (props.item.is_spiked) {
-                        return React.createElement(
-                            'div',
-                            {className: 'expires', key: 'expiry'},
-                            gettext('expires') + ' ' + datetime.shortFormat(props.item.expiry)
-                        );
-                    }
-                },
-
-                desk: function(props) {
-                    if (!props.item.archived) {
-                        return React.createElement(ItemContainer, {item: props.item, desk: props.desk, key: 'desk'});
-                    }
-                },
-
-                fetchedDesk: function(props) {
-                    if (props.item.archived) {
-                        return React.createElement(FetchedDesksInfo, {item: props.item, key: 'desk'});
-                    }
-                },
-
-                priority: function(props) {
-                    return props.item.priority ?
-                        React.createElement(ItemPriority, angular.extend({key: 'priority'}, props.item))
-                        : null;
-                },
-
-                urgency: function(props) {
-                    return props.item.urgency ?
-                        React.createElement(ItemUrgency, angular.extend({key: 'urgency'}, props.item))
-                        : null;
-                }
-            };
-
-            var ActionsMenu = React.createClass({
-                toggle: function(event) {
-                    this.stopEvent(event);
-                    closeActionsMenu();
-                    var icon = ReactDOM.findDOMNode(this)
-                        .getElementsByClassName('icon-dots-vertical')[0];
-
-                    renderToBody(this.renderMenu(), icon);
-                },
-
-                stopEvent: function(event) {
-                    event.stopPropagation();
-                },
-
-                getActions: function() {
-                    var item = this.props.item;
-                    var type = this.getType();
-                    var intent = {action: 'list', type: type};
-                    var groups = {};
-
-                    superdesk.findActivities(intent, item).forEach((activity) => {
-                        if (workflowService.isActionAllowed(item, activity.action)) {
-                            var group = activity.group || 'default';
-
-                            groups[group] = groups[group] || [];
-                            groups[group].push(activity);
-                        }
-                    });
-                    return groups;
-                },
-
-                getType: function() {
-                    return archiveService.getType(this.props.item);
-                },
-
-                groups: [
-                    {_id: 'default', label: gettext('Actions')},
-                    {_id: 'packaging', label: gettext('Packaging')},
-                    {_id: 'highlights', label: gettext('Highlights')},
-                    {_id: 'corrections', label: gettext('Corrections')}
-                ],
-
-                renderMenu: function() {
-                    var menu = [];
-                    var item = this.props.item;
-
-                    var createAction = function(activity) {
-                        return React.createElement(ActionsMenu.Item, {
-                            item: item,
-                            activity: activity,
-                            key: activity._id
-                        });
-                    };
-
-                    var actions = this.getActions();
-
-                    this.groups.map((group) => {
-                        if (actions[group._id]) {
-                            if (group.label === 'Actions') {
-                                menu.push(
-                                    React.createElement(ActionsMenu.Label, {
-                                        label: group.label,
-                                        key: 'group-label-' + group._id
-                                    }),
-                                    React.createElement(ActionsMenu.Divider, {
-                                        key: 'group-divider-' + group._id
-                                    })
-                                );
-                            } else {
-                                menu.push(
-                                    React.createElement(ActionsMenu.Divider, {
-                                        key: 'group-divider-' + group._id
-                                    }),
-                                    React.createElement(ActionsMenu.Label, {
-                                        label: group.label,
-                                        key: 'group-label-' + group._id
-                                    })
-                                );
-                            }
-
-                            menu.push(...actions[group._id].map(createAction));
-                        }
-                        return null;
-                    });
-
-                    return React.createElement(
-                        'ul',
-                        {
-                            className: 'dropdown dropdown__menu more-activity-menu open',
-                            style: {display: 'block', minWidth: 200}
-                        },
-                        menu
-                    );
-                },
-
-                render: function() {
-                    return React.createElement(
-                        'div',
-                        {className: 'item-right toolbox'},
-
-                        React.createElement(
-                            'div',
-                            {className: 'item-actions-menu dropdown--big open'},
-                            React.createElement(
-                                'button',
-                                {
-                                    className: 'more-activity-toggle condensed dropdown__toggle',
-                                    onClick: this.toggle,
-                                    onDoubleClick: this.stopEvent
-                                },
-                                React.createElement('i', {className: 'icon-dots-vertical'})
-                            )
-                        )
-                    );
-                }
-            });
-
-            ActionsMenu.Label = function(props) {
-                return React.createElement(
-                    'li',
-                    null,
-                    React.createElement('div', {
-                        className: 'dropdown__menu-label'
-                    }, gettextCatalog.getString(props.label),
-                        props.label === 'Actions' ? React.createElement(
-                            'button',
-                            {className: 'dropdown__menu-close', onClick: closeActionsMenu},
-                            React.createElement(
-                                'i',
-                                {className: 'icon-close-small'}
-                            )
-                        ) : null
-                    )
-                );
-            };
-
-            ActionsMenu.Divider = function() {
-                return React.createElement('li', {className: 'divider'});
-            };
-
-            ActionsMenu.Item = React.createClass({
-                run: function(event) {
-                    // Stop event propagation so that click on item action
-                    // won't select that item for preview/authoring.
-                    event.stopPropagation();
-                    scope.$apply(() => {
-                        $rootScope.$broadcast('broadcast:preview', {item: null}); // closes preview if already opened
-                        activityService.start(this.props.activity, {data: {item: this.props.item}});
-                    });
-
-                    closeActionsMenu();
-                },
-
-                getInitialState: function() {
-                    return {open: false};
-                },
-
-                open: function() {
-                    $timeout.cancel(this.closeTimeout);
-                    this.closeTimeout = null;
-                    if (!this.state.open) {
-                        this.setPosition();
-                        this.setState({open: true});
-                    }
-                },
-
-                setPosition: function() {
-                    var targetRect = ReactDOM.findDOMNode(this).getBoundingClientRect();
-                    var LEFT_BAR_WIDTH = 48;
-                    var BUFFER = 250;
-
-                    if (targetRect.left < LEFT_BAR_WIDTH + BUFFER) {
-                        this.setState({position: 'left-submenu'});
-                    } else {
-                        this.setState({position: 'right-submenu'});
-                    }
-                },
-
-                close: function() {
-                    if (this.state.open && !this.closeTimeout) {
-                        this.closeTimeout = $timeout(() => {
-                            this.closeTimeout = null;
-                            this.setState({open: false});
-                        }, 100, false);
-                    }
-                },
-
-                closeMenu: function(event) {
-                    // called by the onclick event of the submenu dropdown to close actions menu.
-                    event.stopPropagation();
-                    closeActionsMenu();
-                },
-
-                componentWillUnmount: function() {
-                    $timeout.cancel(this.closeTimeout);
-                    this.closeTimeout = null;
-                },
-
-                render: function() {
-                    var activity = this.props.activity;
-
-                    if (activity.dropdown) {
-                        return React.createElement(
-                            'li',
-                            {onMouseEnter: this.open, onMouseLeave: this.close, onClick: this.closeMenu},
-                            React.createElement(
-                                'div',
-                                {className: 'dropdown dropdown--noarrow' + (this.state.open ? ' open' : '')},
-                                React.createElement(
-                                    'a',
-                                    {
-                                        className: 'dropdown__toggle',
-                                        title: gettextCatalog.getString(activity.label)
-                                    },
-                                    activity.icon ? React.createElement('i', {
-                                        className: 'icon-' + activity.icon
-                                    }, '') : null,
-                                    gettextCatalog.getString(activity.label),
-                                    React.createElement('i', {
-                                        className: 'icon-chevron-right-thin dropdown__submenu-icon'
-                                    })
-                                ),
-                                this.state.open ? $injector.invoke(activity.dropdown, activity, {
-                                    item: this.props.item,
-                                    className: 'dropdown__menu upward ' + this.state.position,
-                                    noHighlightsLabel: gettextCatalog.getString('No available highlights'),
-                                    noLanguagesLabel: gettextCatalog.getString('No available translations')
-                                }) : null
-                            )
-                        );
-                    }
-
-                    return React.createElement(
-                        'li',
-                        null,
-                        React.createElement(
-                            'a',
-                            {title: gettextCatalog.getString(activity.label), onClick: this.run},
-                            React.createElement('i', {
-                                className: 'icon-' + activity.icon
-                            }),
-                            React.createElement('span', {
-                                style: {display: 'inline'}
-                            }, gettextCatalog.getString(activity.label))
-                        )
-                    );
-                }
-            });
-
-            var ProgressBar = function(props) {
-                return React.createElement('div', {
-                    className: 'archiving-progress',
-                    style: {width: props.completed + '%'}
-                });
-            };
-
-            var ErrorBox = function(props) {
-                return React.createElement('div', {className: 'error-box'},
-                    React.createElement('p', {className: 'message'},
-                    gettextCatalog.getString('There was an error archiving this item')),
-                    React.createElement('div', {className: 'buttons'})
-                );
-            };
-
-            /**
-             * Item component
-             */
-            var Item = React.createClass({
-                shouldComponentUpdate: function(nextProps, nextState) {
-                    return nextProps.swimlane !== this.props.swimlane || nextProps.item !== this.props.item ||
-                        nextProps.view !== this.props.view ||
-                        nextProps.flags.selected !== this.props.flags.selected ||
-                        nextState !== this.state;
-                },
-
-                select: function(event) {
-                    if (!this.props.item.gone) {
-                        this.props.onSelect(this.props.item, event);
-                    }
-                },
-
-                selectTakesPackage: function() {
-                    this.props.onSelect(this.props.item.takes || this.props.item.archive_item.takes);
-                },
-
-                selectUpdate: function() {
-                    authoringWorkspace.edit({_id: this.props.item.rewritten_by}, 'view');
-                },
-
-                edit: function(event) {
-                    if (!this.props.item.gone) {
-                        this.props.onEdit(this.props.item);
-                    }
-                },
-
-                dbClick: function(event) {
-                    if (!this.props.item.gone) {
-                        this.props.onDbClick(this.props.item);
-                    }
-                },
-
-                getInitialState: function() {
-                    return {hover: false};
-                },
-
-                setHoverState: function() {
-                    this.setState({hover: true});
-                },
-
-                unsetHoverState: function() {
-                    this.setState({hover: false});
-                },
-
-                onDragStart: function(event) {
-                    dragitem.start(event, this.props.item);
-                },
-
-                render: function() {
-                    var item = this.props.item;
-                    var contents = [
-                        'div',
-                        {
-                            className: classNames(
-                                'media-box',
-                                'media-' + item.type,
-                                {
-                                    locked: item.lock_user && item.lock_session,
-                                    selected: this.props.flags.selected,
-                                    archived: item.archived || item.created,
-                                    gone: item.gone
-                                }
-                            )
-
-                        }
-                    ];
-
-                    if (item._progress) {
-                        contents.push(React.createElement(ProgressBar, {completed: item._progress}));
-                    }
-
-                    const getActionsMenu = () =>
-                        this.state.hover && !item.gone ? React.createElement(ActionsMenu, {item: item}) : null;
-
-                    if (this.props.view === 'mgrid') {
-                        contents.push(
-                            item.archiveError ? React.createElement(ErrorBox) : null,
-                            React.createElement(MediaPreview, {
-                                item: item,
-                                desk: this.props.desk,
-                                onMultiSelect: this.props.onMultiSelect,
-                                swimlane: this.props.swimlane
-                            }),
-                            React.createElement(MediaInfo, {item: item, ingestProvider: this.props.ingestProvider}),
-                            React.createElement(GridTypeIcon, {item: item}),
-                            item.priority ? React.createElement(ItemPriority, item) : null,
-                            item.urgency ? React.createElement(ItemUrgency, item) : null,
-                            fields.broadcast({item: item}),
-                            getActionsMenu()
-                        );
-                    } else {
-                        contents.push(
-                            React.createElement('span', {className: 'state-border'}),
-                            React.createElement(ListTypeIcon, {
-                                item: item,
-                                onMultiSelect: this.props.onMultiSelect,
-                                swimlane: this.props.swimlane
-                            }),
-                            item.priority || item.urgency ? React.createElement(ListPriority, {item: item}) : null,
-                            React.createElement(ListItemInfo, {
-                                item: item,
-                                selectTakesPackage: this.selectTakesPackage,
-                                selectUpdate: this.selectUpdate,
-                                desk: this.props.desk,
-                                ingestProvider: this.props.ingestProvider,
-                                highlightsById: this.props.highlightsById,
-                                profilesById: this.props.profilesById,
-                                swimlane: this.props.swimlane,
-                                versioncreator: this.props.versioncreator
-                            }),
-                            getActionsMenu()
-                        );
-                    }
-
-                    return React.createElement(
-                        'li',
-                        {
-                            id: item._id,
-                            key: item._id,
-                            className: classNames(
-                                'list-item-view',
-                                {active: this.props.flags.selected},
-                                {selected: this.props.item.selected}
-                            ),
-                            onMouseEnter: this.setHoverState,
-                            onMouseLeave: this.unsetHoverState,
-                            onDragStart: this.onDragStart,
-                            onClick: this.select,
-                            onDoubleClick: this.dbClick,
-                            draggable: true,
-                            tabIndex: '0'
-                        },
-                        React.createElement.apply(null, contents)
-                    );
-                }
-            });
-
-            /**
-             * Item list component
-             */
-            var ItemList = React.createClass({
-                getInitialState: function() {
-                    return {itemsList: [], itemsById: {}, selected: null, view: 'mgrid'};
-                },
-
-                multiSelect: function(items, selected) {
-                    var itemsById = angular.extend({}, this.state.itemsById);
-
-                    items.forEach((item) => {
-                        var itemId = search.generateTrackByIdentifier(item);
-
-                        itemsById[itemId] = angular.extend({}, item, {selected: selected});
-                        scope.$applyAsync(() => {
-                            multi.toggle(itemsById[itemId]);
-                        });
-                    });
-
-                    this.setState({itemsById: itemsById});
-                },
-
-                select: function(item, event) {
-                    if (event && event.shiftKey) {
-                        return this.selectMultipleItems(item);
-                    }
-
-                    this.setSelectedItem(item);
-
-                    if (event && event.ctrlKey) {
-                        return this.selectItem(item);
-                    }
-
-                    $timeout.cancel(this.updateTimeout);
-                    this.updateTimeout = $timeout(() => {
-                        if (item && scope.preview) {
-                            scope.$apply(() => {
-                                scope.preview(item);
-                            });
-                        }
-                    }, 500, false);
-                },
-
-                selectItem: function(item) {
-                    if (isCheckAllowed(item)) {
-                        var selected = !item.selected;
-
-                        this.multiSelect([item], selected);
-                    }
-                },
-
-                selectMultipleItems: function(lastItem) {
-                    var itemId = search.generateTrackByIdentifier(lastItem),
-                        positionStart = 0,
-                        positionEnd = _.indexOf(this.state.itemsList, itemId),
-                        selectedItems = [];
-
-                    if (this.state.selected) {
-                        positionStart = _.indexOf(this.state.itemsList, this.state.selected);
-                    }
-
-                    for (var i = positionStart; i <= positionEnd; i++) {
-                        var item = this.state.itemsById[this.state.itemsList[i]];
-
-                        if (isCheckAllowed(item)) {
-                            selectedItems.push(item);
-                        }
-                    }
-
-                    multi.reset();
-                    this.multiSelect(selectedItems, true);
-                },
-
-                dbClick: function(item) {
-                    var activities = superdesk.findActivities({action: 'list', type: 'archive'}, item);
-                    var canEdit = _.reduce(activities, (result, value) => result || value._id === 'edit.item', false);
-
-                    this.setSelectedItem(item);
-                    $timeout.cancel(this.updateTimeout);
-
-                    if (canEdit && scope.edit) {
-                        scope.$apply(() => {
-                            scope.edit(item);
-                        });
-                    } else {
-                        scope.$apply(() => {
-                            authoringWorkspace.open(item);
-                        });
-                    }
-                },
-
-                edit: function(item) {
-                    this.setSelectedItem(item);
-                    $timeout.cancel(this.updateTimeout);
-                    if (item && scope.edit) {
-                        scope.$apply(() => {
-                            scope.edit(item);
-                        });
-                    } else if (item) {
-                        scope.$apply(() => {
-                            authoringWorkspace.open(item);
-                        });
-                    }
-                },
-
-                deselectAll: function() {
-                    this.setState({selected: null});
-                },
-
-                updateAllItems: function(itemId, changes) {
-                    var itemsById = angular.extend({}, this.state.itemsById);
-
-                    _.forOwn(itemsById, (value, key) => {
-                        if (_.startsWith(key, itemId)) {
-                            itemsById[key] = angular.extend({}, value, changes);
-                        }
-                    });
-
-                    this.setState({itemsById: itemsById});
-                },
-
-                findItemByPrefix: function(prefix) {
-                    var item;
-
-                    _.forOwn(this.state.itemsById, (val, key) => {
-                        if (_.startsWith(key, prefix)) {
-                            item = val;
-                        }
-                    });
-
-                    return item;
-                },
-
-                setSelectedItem: function(item) {
-                    if (monitoringState.selectedGroup !== scope.$id) {
-                        // If selected item is from another group, deselect all
-                        $rootScope.$broadcast('item:unselect');
-                        monitoringState.selectedGroup = scope.$id;
-                    }
-
-                    this.setState({selected: item ? search.generateTrackByIdentifier(item) : null});
-                },
-
-                getSelectedItem: function() {
-                    var selected = this.state.selected;
-
-                    return this.state.itemsById[selected];
-                },
-
-                updateItem: function(itemId, changes) {
-                    var item = this.state.itemsById[itemId] || null;
-
-                    if (item) {
-                        var itemsById = angular.extend({}, this.state.itemsById);
-
-                        itemsById[itemId] = angular.extend({}, item, changes);
-                        this.setState({itemsById: itemsById});
-                    }
-                },
-
-                handleKey: function(event) {
-                    var diff;
-
-                    switch (event.keyCode) {
-                    case Keys.right:
-                    case Keys.down:
-                        diff = 1;
-                        break;
-
-                    case Keys.left:
-                    case Keys.up:
-                        diff = -1;
-                        break;
-
-                    case Keys.enter:
-                        if (this.state.selected) {
-                            this.edit(this.getSelectedItem());
-                        }
-
-                        event.stopPropagation();
-                        return;
-
-                    case Keys.pageup:
-                    case Keys.pagedown:
-                        event.preventDefault();
-                        event.stopPropagation();
-                        this.select(); // deselect active item
-                        scope.$applyAsync(() => {
-                            monitoringState.moveActiveGroup(event.keyCode === Keys.pageup ? -1 : 1);
-                        });
-                        return;
-                    }
-
-                    var highlightSelected = () => {
-                        for (var i = 0; i < this.state.itemsList.length; i++) {
-                            if (this.state.itemsList[i] === this.state.selected) {
-                                var next = Math.min(this.state.itemsList.length - 1, Math.max(0, i + diff));
-
-                                this.select(this.state.itemsById[this.state.itemsList[next]]);
-                                return;
-                            }
-                        }
-                    };
-
-                    const checkRemaining = () => {
-                        if (!_.isNil(diff)) {
-                            event.preventDefault();
-                            event.stopPropagation();
-
-                            if (this.state.selected) {
-                                highlightSelected();
-                            } else {
-                                this.select(this.state.itemsById[this.state.itemsList[0]]);
-                            }
-                        }
-                    };
-
-                    checkRemaining();
-                },
-
-                closeActionsMenu: closeActionsMenu,
-
-                componentWillUnmount: function() {
-                    this.closeActionsMenu();
-                },
-
-                componentWillUpdate: function() {
-                    this.closeActionsMenu();
-                },
-
-                setSelectedComponent: function(com) {
-                    this.selectedCom = com;
-                },
-
-                modifiedUserName: function(versionCreator) {
-                    return this.props.usersById[versionCreator] ?
-                        this.props.usersById[versionCreator].display_name : null;
-                },
-
-                render: function render() {
-                    var createItem = function createItem(itemId) {
-                        var item = this.state.itemsById[itemId];
-                        var task = item.task || {desk: null};
-
-                        return React.createElement(Item, {
-                            key: itemId,
-                            item: item,
-                            view: this.state.view,
-                            swimlane: this.state.swimlane || storage.getItem('displaySwimlane'),
-                            flags: {selected: this.state.selected === itemId},
-                            onEdit: this.edit,
-                            onDbClick: this.dbClick,
-                            onSelect: this.select,
-                            onMultiSelect: this.multiSelect,
-                            ingestProvider: this.props.ingestProvidersById[item.ingest_provider] || null,
-                            desk: this.props.desksById[task.desk] || null,
-                            highlightsById: this.props.highlightsById,
-                            profilesById: this.props.profilesById,
-                            setSelectedComponent: this.setSelectedComponent,
-                            versioncreator: this.modifiedUserName(item.version_creator)
-                        });
-                    }.bind(this);
-                    var isEmpty = !this.state.itemsList.length;
-
-                    return React.createElement(
-                        'ul',
-                        {
-                            className: classNames(
-                                'list-view',
-                                this.state.view + '-view',
-                                {'list-without-items': isEmpty}
-                            ),
-                            onClick: this.closeActionsMenu
-                        },
-                        isEmpty && !scope.loading ?
-                            React.createElement(
-                                'li',
-                                {onClick: this.closeActionsMenu},
-                                gettextCatalog.getString('There are currently no items')
-                            ) : this.state.itemsList.map(createItem)
-                    );
-                }
-            });
-
             monitoringState.init().then(() => {
-                var itemList = React.createElement(ItemList, monitoringState.state);
+                var itemList = React.createElement(ItemListComponent,
+                    angular.extend({
+                        svc: services,
+                        scope: scope
+                    }, monitoringState.state));
+
                 var listComponent = ReactDOM.render(itemList, elem[0]);
 
                 /**
